requirements.txt
django
djangorestframework
django-rest-passwordreset
psycopg2
Pillow